# dev: reg: areg: Speed up reg and areg by only aligning to first 1000 items.