--- a/hledger/Hledger/Cli/Commands/Register.hs
+++ b/hledger/Hledger/Cli/Commands/Register.hs
@@ -96,8 +96,8 @@ postingsReportAsText :: CliOpts -> PostingsReport -> TL.Text
 postingsReportAsText opts items = TB.toLazyText lines
   where
     lines = foldMap (postingsReportItemAsText opts amtwidth balwidth) items
-    amtwidth = maximumStrict $ 12 : widths (map itemamt items)
-    balwidth = maximumStrict $ 12 : widths (map itembal items)
+    amtwidth = maximumStrict $ 12 : widths (map itemamt $ take 1000 items)
+    balwidth = maximumStrict $ 12 : widths (map itembal $ take 1000 items)
     widths = map wbWidth . concatMap (showMixedAmountLinesB oneLine)
     itemamt (_,_,_,Posting{pamount=a},_) = a
     itembal (_,_,_,_,a) = a