@Xitian9 2022-03-04
This is a simple trick which basically just puts a limit on the number of postings we need to hold in memory at the same time. The price is that we lose the ability to align to any long postings which occur beyond the 1000th posting in a register or aregister report. There is no change at all for reports with fewer than 1000 postings.

How often we want such long reports, and how much it's worth optimising around them, is a judgement call. We could also change 1000 to any other positive number.